+++ CHANGELOG.rst
@@ -12,7 +12,8 @@ Changelog
 0.17.3
 ------
 - Fix duplicates when using custom through association class on M2M relations
-- Fix `update_or_create`. (#721)
+- Fix `update_or_create` and `get_or_create`. (#721)
+- Fix `refresh_from_db` without fields pass. (#734)
 
 0.17.2
 ------